move path tests to a separate package
